+++ jslint_ci.sh
@@ -1062,7 +1062,7 @@ import moduleFs from "fs";
     default:
         version = `p${version}`;
         commitMessage = (
-            /\n\n# v\d\d\d\d\.\d\d?\.\d\d?(?:-.*?)?\n(- [\S\s]+?)\n- /
+            /\n\n# v\d\d\d\d\.\d\d?\.\d\d?(?:-.*?)?\n(- [\S\s]+?)(?:\n- |\n\n)/
         ).exec(data)[1];
     }
     branchPull = `branch-${version}`;
@@ -2068,7 +2068,7 @@ function replaceListReplace(replaceList, data) {
             if (dataUriType) {
                 return;
             }
-            if (dateCommitted) {
+            if (dateCommitted && dateCommitted.toString()) {
                 result += (
                     "\n\n\n/*\n"
                     + "repo " + prefix.replace("/blob/", "/tree/") + "\n"
@@ -3329,20 +3329,27 @@ function sentinel() {}
       }
     }));
     exitCode = await new Promise(function (resolve) {
+      let processArgv0 = processArgv[0];
+      if (processArgv0 === "npm") {
+        processArgv0 = process.platform.replace(
+          "win32",
+          "npm.cmd"
+        ).replace(
+          process.platform,
+          "npm"
+        );
+      }
       moduleChildProcess.spawn(
-        (
-          processArgv[0] === "npm"
-          ? process.platform.replace("win32", "npm.cmd").replace(
-            process.platform,
-            "npm"
-          )
-          : processArgv[0]
-        ),
+        processArgv0,
         processArgv.slice(1),
         {
           env: Object.assign({}, process.env, {
             NODE_V8_COVERAGE: coverageDir
           }),
+          shell: (
+            processArgv0.endsWith(".bat")
+            || processArgv0.endsWith(".cmd")
+          ),
           stdio: ["ignore", 1, 2]
         }
       ).on("exit", resolve);